requirements.txt
psycopg2~=2.9.9
python-decouple~=3.8